ajout 1-string_nconcat.c

+++ 0x0C-more_malloc_free/1-string_nconcat.c
@@ -12,7 +12,7 @@
 char *string_nconcat(char *s1, char *s2, unsigned int n)
 {
     char *string;
-    unsigned long long i = 0, j = 0, len_s1 = 0, len_s2 = 0, len_string = 0;
+    unsigned int i = 0, j = 0, len_s1 = 0, len_s2 = 0, len_string = 0;
 
     while (s1 && s1[len_s1])
         len_s1++;
